requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
mysql-connector-python
SQLAlchemy
mysqlclient
bcrypt